feat: adds loading for the whole registration entry
