+++ src/core_modules/capture-core/components/DataEntries/EnrollmentRegistrationEntry/EnrollmentRegistrationEntry.component.js
@@ -1,6 +1,7 @@
 // @flow
 import React, { type ComponentType } from 'react';
 import { Button } from '@dhis2/ui';
+import i18n from '@dhis2/d2-i18n';
 import { withStyles } from '@material-ui/core';
 import { compose } from 'redux';
 import { useScopeInfo } from '../../../hooks/useScopeInfo';
@@ -8,14 +9,23 @@ import { scopeTypes } from '../../../metaData';
 import { EnrollmentDataEntry } from '../Enrollment';
 import { useCurrentOrgUnitInfo } from '../../../hooks/useCurrentOrgUnitInfo';
 import { useRegistrationFormInfoForSelectedScope } from '../common/useRegistrationFormInfoForSelectedScope';
-import type { OwnProps } from './EnrollmentRegistrationEntry.types';
+import type { OwnProps, Props } from './EnrollmentRegistrationEntry.types';
 import { withSaveHandler } from '../../DataEntry';
+import { withLoadingIndicator } from '../../../HOC';
+import { InfoIconText } from '../../InfoIconText';
 
 const styles = ({ typography }) => ({
     marginTop: {
         marginTop: typography.pxToRem(2),
     },
 });
+
+const translatedTextWithStylesForProgram = (trackedEntityName: string, programName: string, orgUnitName: string) =>
+    (<>
+        {i18n.t('Saving a {{trackedEntityName}} in', { trackedEntityName })} <b>{programName}</b>
+        {orgUnitName && <>{' '}{i18n.t('in')} <b>{orgUnitName}</b></>}.
+    </>);
+
 
 const EnrollmentRegistrationEntryPlain =
   ({
@@ -26,8 +36,8 @@ const EnrollmentRegistrationEntryPlain =
       classes,
       onSave,
       ...rest
-  }: {...OwnProps, ...CssClasses}) => {
-      const { scopeType } = useScopeInfo(selectedScopeId);
+  }: Props) => {
+      const { scopeType, trackedEntityName, programName } = useScopeInfo(selectedScopeId);
       const { formId, formFoundation } = useRegistrationFormInfoForSelectedScope(selectedScopeId);
       const orgUnit = useCurrentOrgUnitInfo();
 
@@ -57,14 +67,18 @@ const EnrollmentRegistrationEntryPlain =
                           </Button>
                       }
 
+                      <InfoIconText
+                          text={translatedTextWithStylesForProgram(trackedEntityName.toLowerCase(), programName, orgUnit.name)}
+                      />
                   </>
               }
           </>
       );
   };
 
-export const EnrollmentRegistrationEntryComponent: ComponentType<OwnProps> =
+export const EnrollmentRegistrationEntryComponent: ComponentType<$Diff<Props, CssClasses>> =
   compose(
+      withLoadingIndicator(() => ({ height: '350px' })),
       withSaveHandler({ onGetFormFoundation: ({ enrollmentMetadata }) => enrollmentMetadata && enrollmentMetadata.enrollmentForm }),
       withStyles(styles),
   )(EnrollmentRegistrationEntryPlain);
