+++ src/core_modules/capture-core/components/DataEntries/TeiRegistrationEntry/TeiRegistrationEntry.types.js
@@ -11,3 +11,14 @@ export type OwnProps = $ReadOnly<{|
   onGetUnsavedAttributeValues: Function,
   onPostProcessErrorMessage: Function,
 |}>
+
+type ContainerProps = {|
+  ready: boolean,
+|}
+
+export type Props = $ReadOnly<{|
+  ...OwnProps,
+  ...CssClasses,
+  ...ContainerProps
+|}>
+
