+++ src/core_modules/capture-core/components/DataEntry/DataEntry.container.js
@@ -43,5 +43,5 @@ const mapDispatchToProps = (dispatch: ReduxDispatch) => ({
 export default compose(
     // $FlowFixMe
     connect(mapStateToProps, mapDispatchToProps, null),
-    withLoadingIndicator(() => ({ height: '350px' }), null),
+    withLoadingIndicator(() => ({ height: '350px' })),
 )(DataEntry);
