+++ src/core_modules/capture-core/components/Pages/New/RegistrationDataEntry/RegistrationDataEntry.component.js
@@ -1,6 +1,5 @@
 // @flow
 import React from 'react';
-import i18n from '@dhis2/d2-i18n';
 import Paper from '@material-ui/core/Paper/Paper';
 import { withStyles } from '@material-ui/core';
 import type { Props } from './RegistrationDataEntry.types';
@@ -8,25 +7,10 @@ import { EnrollmentRegistrationEntry, TeiRegistrationEntry, SingleEventRegistrat
 import { scopeTypes } from '../../../../metaData';
 import { useScopeInfo } from '../../../../hooks/useScopeInfo';
 import { useRegistrationFormInfoForSelectedScope } from '../../../DataEntries/common/useRegistrationFormInfoForSelectedScope';
-import { InfoIconText } from '../../../InfoIconText';
 import { useCurrentOrgUnitInfo } from '../../../../hooks/useCurrentOrgUnitInfo';
 import { useScopeTitleText } from '../../../../hooks/useScopeTitleText';
 import { TrackedEntityTypeSelector } from '../../../TrackedEntityTypeSelector';
 
-
-const translatedTextWithStylesForProgram = (trackedEntityName: string, programName: string, orgUnitName: string) =>
-    (<>
-        {i18n.t('Saving a {{trackedEntityName}} in', { trackedEntityName })} <b>{programName}</b>
-        {orgUnitName && <>{' '}{i18n.t('in')} <b>{orgUnitName}</b></>}.
-    </>);
-
-
-const translatedTextWithStylesForTei = (trackedEntityName, orgUnitName) =>
-    (<>
-        {i18n.t('Saving a {{trackedEntityName}}', { trackedEntityName })} <b>{i18n.t('without')}</b> {i18n.t('enrollment')}
-        {orgUnitName && <>{' '}{i18n.t('in')} <b>{orgUnitName}</b></>}.{' '}
-        {i18n.t('Enroll in a program by selecting a program from the top bar.')}
-    </>);
 
 const getStyles = ({ typography }) => ({
     paper: {
@@ -55,9 +39,8 @@ const RegistrationDataEntryPlain = (
         onSaveWithoutEnrollment,
         onSaveWithEnrollment,
     }: Props) => {
-    const { scopeType, trackedEntityName, programName } = useScopeInfo(selectedScopeId);
+    const { scopeType } = useScopeInfo(selectedScopeId);
     const { registrationMetaData } = useRegistrationFormInfoForSelectedScope(selectedScopeId);
-    const orgUnit = useCurrentOrgUnitInfo();
     const titleText = useScopeTitleText(selectedScopeId);
 
     const handleRegistrationScopeSelection = (id) => {
@@ -97,9 +80,6 @@ const RegistrationDataEntryPlain = (
                             onUpdateField={() => console.log('onUpdateField will be here in the future in the future')}
                             onStartAsyncUpdateField={() => console.log('onStartAsyncUpdateField will be here in the future in the future')}
                         />
-                        <InfoIconText
-                            text={translatedTextWithStylesForProgram(trackedEntityName.toLowerCase(), programName, orgUnit.name)}
-                        />
                     </div>
                 </Paper>
             }
@@ -115,7 +95,6 @@ const RegistrationDataEntryPlain = (
                         <TrackedEntityTypeSelector onSelect={handleRegistrationScopeSelection} />
                     </div>
                     <div className={classes.registrationContainer}>
-
                         <TeiRegistrationEntry
                             id={dataEntryId}
                             selectedScopeId={selectedScopeId}
@@ -124,9 +103,6 @@ const RegistrationDataEntryPlain = (
                             onSave={onSaveWithoutEnrollment}
                             onGetUnsavedAttributeValues={() => console.log('onGetUnsavedAttributeValues will be here in the future in the future')}
                             onPostProcessErrorMessage={() => console.log('onPostProcessErrorMessage will be here in the future in the future')}
-                        />
-                        <InfoIconText
-                            text={translatedTextWithStylesForTei(trackedEntityName.toLowerCase(), orgUnit.name)}
                         />
                     </div>
                 </Paper>
