+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/GeneralOutput/GeneralOutput.container.js
@@ -4,11 +4,12 @@ import React from 'react';
 import GeneralOutput from './GeneralOutput.component';
 import getDataEntryKey from '../../../../DataEntry/common/getDataEntryKey';
 
-const mapStateToProps = (state: ReduxState) => {
+const mapStateToProps = (state: ReduxState, { id }) => {
     const registerTeiContainer = state.newRelationshipRegisterTei;
     const ready = !registerTeiContainer.loading;
 
-    const dataEntryId = 'relationship';
+    const dataEntryId = id || 'relationship';
+    debugger;
     const dataEntryKey = ready ? getDataEntryKey(dataEntryId, state.dataEntries[dataEntryId].itemId) : null;
     return {
         ready,
@@ -16,7 +17,6 @@ const mapStateToProps = (state: ReduxState) => {
     };
 };
 
-// $FlowSuppress
 // $FlowFixMe[missing-annot] automated comment
 export default connect(mapStateToProps, () => ({}))(
     (props: Object) => (props.ready ? <GeneralOutput {...props} /> : null),
